requirements_telethon.txt
flask>=2.3.0
telethon>=1.24.0
python-dotenv>=1.0.0
